requirements.txt
openai-whisper
ffmpeg-python
torch
pytube
